Adding the feature to create AcordeMenor
