+++ src/main/java/cl/bodoque/tones/model/Acorde.java
@@ -1,25 +1,9 @@
 package cl.bodoque.tones.model;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import lombok.Getter;
+public interface Acorde {
+    public Tono getTonica();
+    public List<Tono> getTonos();
 
-@Getter
-public class Acorde {
-    private List<Tono> tonos;
-    
-    public Acorde(Tono tonica, List<Double> offsets) {
-        tonos = new ArrayList<Tono>(offsets.size() + 1);
-        tonos.add(tonica);
-        Tono actual = tonica;
-        for(Double offset: offsets) {
-            actual = actual.masTonos(offset).orElseThrow(() -> new IllegalStateException("No hay tono"));
-            tonos.add(actual);
-        }
-    }
-    
-    public Tono getTonica() {
-        return tonos.get(0);
-    }
 }
