+++ src/main/java/cl/bodoque/tones/model/AcordeMayor.java
@@ -2,7 +2,7 @@ package cl.bodoque.tones.model;
 
 import java.util.List;
 
-public class AcordeMayor extends Acorde {
+public class AcordeMayor extends AcordeAbstracto {
     public static final List<Double> OFFSETS = List.of(2d, 1.5);
     
     public AcordeMayor(Tono tonica) {
