+++ src/main/java/cl/bodoque/tones/model/AcordeMenor.java
@@ -1,0 +1,11 @@
+package cl.bodoque.tones.model;
+
+import java.util.List;
+
+public class AcordeMenor extends AcordeAbstracto {
+    public static final List<Double> OFFSETS = List.of(1.5, 2d);
+
+    public AcordeMenor(Tono tonica) {
+        super(tonica, OFFSETS);
+    }
+}
